fix(settings): allow settings for ENVIRONMENT
- move analytics.client.timeout in portal config only
Closes gravitee-io/issues#4811
