--- gravitee-rest-api-portal/gravitee-rest-api-portal-rest/src/test/resources/io/gravitee/rest/api/portal/rest/mapper/consoleConfigEntity.json
@@ -1,5 +1,0 @@
-{
-  "analytics" : {
-    "clientTimeout" : 30000
-  }
-}
