+++ gravitee-rest-api-portal/gravitee-rest-api-portal-rest/src/test/resources/io/gravitee/rest/api/portal/rest/mapper/portalConfigEntity.json
@@ -1,4 +1,7 @@
 {
+  "analytics" : {
+    "clientTimeout" : 30000
+  },
   "api": {
     "labelsDictionary": [
       "label1",
